+++ phosphor-power-supply/power_supply.cpp
@@ -25,7 +25,7 @@ constexpr auto bindDelay = 1000;
 using namespace sdbusplus::xyz::openbmc_project::Common::Device::Error;
 
 PowerSupply::PowerSupply(
-    sdbusplus::bus_t& bus, const std::string& invpath, std::uint8_t i2cbus,
+    sdbusplus::bus_t& bus, const std::string& invpath, std::uint16_t i2cbus,
     std::uint16_t i2caddr, const std::string& driver,
     const std::string& gpioLineName, std::function<bool()>&& callback) :
     bus(bus), inventoryPath(invpath),
